+++ ios/Classes/WorkoutReading/UserDefaultsKeys.swift
@@ -1,9 +1,6 @@
 import Foundation
 
 public struct UserDefaultsKeys {
-    public static let isImportRunning = "isImportRunning"
-    public static let isImportCycling = "isImportCycling"
-    public static let isImportSwimming = "isImportSwimming"
     public static let apiURL = "apiURL"
     public static let token = "token"
 }